Drop libgcc_post_upgrade utility in favour of lua %post and %postun scripts.
The only reason for libgcc_post_upgrade utility's existence is to
exec /sbin/ldconfig if it's present and return zero exit code in any case.

This was needed in %post script of libgcc package because it's one
of the first packages being installed when creating sandboxed buildroot
or installing a new system, when neither bash nor glibc is available.

The libgcc_post_upgrade program is an architecture-dependent hack,
written using asm statements to make system calls directly,
without relying on glibc.
It's not available on some of architectures supported by Mer: mips, aarch64.

Now rpm has builtin lua interpreter, so this hack is not needed anymore.
This is exactly the way it's done in Fedora, so I've just copied
the %post and %postun scripts for libgcc from Fedora package.

Signed-off-by: Oleg Girko <oleg.girko@jolla.com>
Change-Id: I2f283fbb5faad43f9b1d8f10fb275596315c7574